Part 3: Variables
Creating Variables and Variable Types in Lua
This video explains the basic types of variables in Lua and
how they behave, and in what ways they differ from other languages.
I left all the variable assignments and print statements in the file,
since that makes it clearer what exactly has been done in this video.