Part 5: Math
How to do Math and use the Math Library in Lua to make your life easier
This video shows how to use the common math functions in Lua. An odd
quirk I noticed was that the math.random() functions, was not very
random at all. The use of math.randomseed(os.time()) did affect the
value, but did not make it random. Some Googling resulted in the
advice to call math.random() a couple of times, that did affect the
resulting value.