Part 7: Scope
What is Variable Scope and why is it important?
This video briefly explains about scope in Lua. By default everything
exists in the global scope (variables and functions). By prefixing
variables with "_G." it is forced to be a globally scoped variable,
this is also the default. When preceding a variable or a function
with the keyword "local" it is scoped to where it is defined (i.e.
inside the file or function) and therefore not accesible outside it.